Update biblepay_zh_CN.ts and biblepay_zh_TW.ts
biblepay_zh_CN.ts: Based on the translation for dash and also refer to the translation for bitcoin. The quality of original translation is very poor so I have to do massive work to improve the it.

- Update incorrect translations
- Fix typos
- Fix inconsistency issue
- Fix punctuation issue
- Fix some obvious errors

biblepay_zh_TW.ts: Based on the translation for dash. The quality of original translation is quite good so I just fix some obvious errors.