1.1.5.7-Mandatory Upgrade (at block 77000)
- Adding Attachments: In QT | Send BBP | Click Attach File: Add IPFS
Attached Document.  This allows the user to upload a file with a
transaction, share on IPFS,
and a recipient to receive the document.
In RPC: Added 'exec ipfsadd document'
- In QT Transaction List | Double click a transaction | View Description
Dialog: IPFS Download Document Button (Allows opening an attachment),
IPFS View Document link, and AltLink
Added icon to txlist for tx's with IPFS attachments

- Added How To install IPFS: InstallingIPFS_Ubuntu64.md
- Added exec ipfsget hash filename (download an IPFS file), exec
ipfslist (hosted hashes in biblepay)
- Added exec ipfsquality - The report of Sancs who are ENABLED and
corresponding IPFS quality levels
- Added ability for Sancs to check each others IPFS health levels

- Added spork to support no PODC team requirement, and spork for PODC
team blacklisting, and spork for non-biblepay-team payment percentage
- Fixed the root cause of prior fork (Removed fMining flag) and
prevented miner log-spam, Modified checkblock consensus to be stricter
to help
prevent future forks

- Added future Sanc consensus reward for IPFS (hosted_file_percent *
sanc_uptime * reward) achieved through voting (this is OFF in prod)
- Added dedicated IPFS receive address so we can track total IPFS fees
- Added spork for signed prayer enforcement (potentially reducing
graffiti)
- Refactor sanctuary magnitude assessment process
- Added IPFS fee disclosure to Fee Validation dialog, charge a specific
fee per KB uploaded (set by spork)
- Added Contact Add page (this also handles View, Edit and Delete),
allows the user to store a contact associating a public key with an
optional e-mail and long/lat (this is for the future
Christian camp, data is stored in IPFS, and is entirely optional, we
support GDPR/Delete)
- Added business object serialization to serialize objects into IPFS,
and serialize/deserialize in JSON
- Added RSA file and string encryption (for symmetric business object
serialization and user permissions)
- Added ability to send BBP via e-mail (exec emailbbp e-mailaddress
amount), and phase 1 of Church Tithing
- Change default privatesend amounts *1000, fix milliBBP display issue
- Added BusinessObject List, View and Navigate Grid
- Added GospelLink Add (exec addgospellink url notes)
- Added ability to store a businessobject schema in IPFS and respect the
schema
- BOINC: Added Ease-of-Use for New-Users buttons in QT DC Page: "Run R@H
Diagnostics" Button: Detects if Boinc is installed, if account exists,
if project attached, generates
report with suggested fixes. (This report also displays a warning if
user UTXO is low).
A mirror of this RPC added: 'exec rosettadiagnostics', displays info
textually.
- Added QT DC Button 'Create and Attach R@H Account': This button will
automatically create a RAH account (if one does not exist) and sign up,
attach the project (if not
attached), start the tasks, and update the display.  A mirror of this is
now in the RPC:  'exec attachrosetta user pass'.
- Added Rosetta task count to DC page UI.