1.4.1.7 - Mandatory Upgrade for TestNet
- Fix bug showing old proposals in Proposals List page
- Change sanctuary lockup collateral to 4,500,001
- Upgrade persisted data storage system from v1.0 to v1.1
- Add txlist double click drill in: display of biblepay-classic fields:
height, time, difficulty
- Add RPC leaderboard command (leaderboard [me=true/false] [height ||
last || future]
- Add exec upgradesanc sancname.  This is a BiblePay utility that allows
you to upgrade a non-deterministic sanc to a deterministic sanc.  We
also add the deterministic.conf file, so the user can write scripts to
read their own deterministic sanc keys and automate the deployment of
the keys to the remote server(s).
- Fix low-abn-weight bug
- Make exec join healing display the wiki URL so the user can learn
about the healing campaign first