1.4.4.7 - Leisure Upgrade
- Fix OneClickMining initial crlf issue (Meaning that biblepay.conf will
not be corrupted if user starts with one-click-mining config)
- Add Payment Option 4 for Cameroon-One, and fix BIO URL
- Apollon Integration for Deterministic Sancs:  Add 'exec blscommand
mnppriv blsprivkey'.  This allows the user to set the deterministic
masternodeblsprivkey in the sanc biblepay.conf file.  NOTE: The
EncryptAES256 and DecryptAES256 were added to support this feature (this
is so no other node can discover your mnprivkey or your mnblsprivkey
during transmission or initialization).
- Update Health feature for Sanctuaries only:  Set healthupdate=1 to
enable, healthupdate=0 to disable, default is Enabled on Sanctuaries
only.  If this is enabled, you will see your Sanctuaries health status
(IE Best Block Hash, GSC voting level) in the pool.  This will be
valuable when we host with Apollon.
- Disconnect nodes who ask for too much gov data
- Fix governance.dat bloated file size, on boot.  Reindex governance.dat
on boot if > 25 megs.
- Clean up GSC contract data after 1 day.  Sanctuary Quorums:  Vote to
delete duplicate contracts.
- Decrease memory consumption (by purging gobject cache)
- Limit network bandwidth (by removing unecessary and expired govobjs)